pyproject.toml
[project]
name = "neural-reels"
version = "0.1.0"
description = "AI-Powered Short-Form Video Asset Generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.2.1",
    "moviepy>=2.2.1",
    "pillow>=11.2.1",
    "ruff>=0.11.12",
]
